full refactor of code using more functions
