--- .vscode/settings.json
@@ -1,7 +1,0 @@
-{
-    "workbench.colorCustomizations": {
-        "activityBar.background": "#4A2213",
-        "titleBar.activeBackground": "#68301B",
-        "titleBar.activeForeground": "#FDF9F7"
-    }
-}
