+++ util/hub_eval/hub_eval/charts_lib.py
@@ -7,7 +7,8 @@ import subprocess
 import time
 from datetime import timedelta
 
-
+import unicodedata
+import re
 
 
 def get_chart_name(url):
@@ -30,13 +31,6 @@ def count_in_file(str, filename):
                 n += 1
     return n
 
-def evaluate_badrobot(template, log_badrobot):
-    os.system("badrobot scan " + template + " > " + log_badrobot )
-    os.system("cat " + log_badrobot + " | jq '[.[].score] | add' > " + log_badrobot + "_sum")
-    with open(log_badrobot+"_sum") as f:
-        score_badrobot = f.readline().strip('\n')
-    return score_badrobot
-
 def has_evaluation(chart_dict, eval_key):
     if eval_key in chart_dict and "score" in chart_dict[eval_key] and chart_dict[eval_key]["score"] != "":
         return True
@@ -47,3 +41,32 @@ def is_chart_error(chart_dict):
         return True
     return False
 
+
+def needs_update(repo, chart, dic_chart, charts_pss):
+    key = repo + "__" + chart
+    version = dic_chart["version"]
+    if key not in charts_pss:
+        return True
+    if not "status" in charts_pss[key] or "chart_version" not in charts_pss[key]["status"] or charts_pss[key]["status"]["chart_version"] != version:
+        return True
+    return False
+
+def needs_evaluation(repo, chart, tool, charts_db):
+    key = repo + "__" + chart
+    if key not in charts_db:
+        print("  **Error, key %s not found in charts db" % key)
+        return False
+    if "status" not in charts_db[key] or "chart_version" not in charts_db[key]["status"]:
+        print("  Status with version not found in charts db")
+        return False
+    if not "cache" in charts_db[key]["status"] or charts_db[key]["status"]["cache"] != "generated":
+        print("  Chart not generated in charts db, status=%s" % charts_db[key]["status"]["cache"])
+        return False
+    if tool not in charts_db[key] or "chart_version" not in charts_db[key][tool]:
+        print("  Tool %s not in status in charts db" % tool)
+        return True
+    if charts_db[key][tool]["chart_version"] != charts_db[key]["status"]["chart_version"]:
+        print("  Chart version for tool not in charts db")
+        return True
+    print("  Tool %s up to date" % tool)
+    return False
